feat(ErrorMessage): add universal error handling
Errors are now handled in a standardized, universal way for all three tools.
A shared ErrorMessage component displays the error, and each tool manages
a single piece of error state at its top level, to make it easy to manage
errors when users trigger multiple errors, even between different methods
of file retrieval (e.g. as opposed to managing error state as a FIFO or list
data structure).  The error state and set method are simply passed as arguments
to the useFileFetcher and useFileUploader hooks where they are used at the top
level of each tool, and then all of the various issues that can trigger errors
germane to each tool can easily set custom error messages.  If a setError/onError
method is not passed to the hooks, they will default to the original (somewhat
crude, tbh) method of displaying errors via alert().