refactor(FileDropzone): remove unnecessary useCallbacks
Removed 3 unnecessary uses of useCallback. (useCallback only provides a benefit
when 1. dependencies are specified in the dependency array, and/or 2. the function
wrapped in useCallback is being passed as a prop to another component, in order to
provide either referential equality across renders, or to memoize expensive
calculations—none of which is the case for these three simple handlers that belong
solely to this component, require no dependencies, and do not perform expensive
calculations.  The fourth usage of useCallback, however, makes sense.  Unnecessary
usage of useCallback adds unnecessary overhead and overcomplicates code.

Also slightly improved the styling of FileDropzone—added an "upload" icon as in the
regular/non-dnd UploadBox component for consistency, and standardized text colors.