feat: standardize display + add preview scale
The structure, styles, and size of all preview images for each of the three tools
has been standardized.  More importantly, whereas previously images larger than
the maximum container width would be scaled to fit, but the user had no idea what
that scale was (and border radius in the corner rounder tool was not actually being
scaled in the preview according to the real dimensions of the image), now, if an
uploaded (or fetched) image is larger than the maximum container dimensions, a
preview scale will appear above the preview image showing the user exactly at what
scale they are currently viewing the preview image.  This preview is dynamic and
connected to a ResizeObserver, so as the user changes their window size (and this
is particularly important for mobile users and users with smaller window sizes,
where nearly all uploaded images must be scaled), the preview scale will change
dynamically as well.