feat: add custom select comp for mobile opt selectors
At smaller screen sizes, the regular option selector (plus svg scale selector and
border radius selector) component can be too wide to fit on the screen and cannot
resize or wrap to fit.  This commit implements a custom select component that is
swapped for the linear scale selector below the "xs" breakpoint (25rem, the width
of the upload box + fetch from url form, plus 2rem inline padding on either side
= 29rem).  The custom select element is fully accessible and keyboard navigable
and implements its keyboard functionality with the custom useKeydown hook.
Conditional rendering across the "xs" breakpoint is handled with a custom
useMediaQuery hook powered by React's new useSyncExternalStore, which is meant to
interact with browser APIs like this.  When the Select component is rendered instead
of the regular option selector components and the "custom" value is selected the
custom value input is rendered in a row beside the custom Select component, instead
of in a column layout as with the regular option selector.  For the SVG scale
selector, a new value of 0.5 or 1/2 scale was added as the smallest scale, since it
is an extremely common use case for users to halve the size of an image.