refactor(Icons): consolidate + add app icons
All app icons have been consolidated into a single shared component file with multiple exported
icon components, each sharing common default props, and exposing a handful of props like strokeWidth
and className. In refactoring icons across the app in this manner, I have also added icons to a few
important places.  A download icon has been added to the main "save" buttons in each of the three
tools, which really helps to communicate the fact that those buttons are primary CTAs and highlights
their action in an immediate, visual way, and a GitHub logo icon has been added to the footer link
to the app's repo on GitHub. This helps balance and standardize the GitHub link in the footer with
the "Back" link in the header. with its icon.