feat(SvgTool): add preview background color
Previously, if users uploaded a transparent SVG with dark (esp. black) artwork,
the image would be completely unreadable on the site's dark/black background.

This commit adds an option selector to the svg-to-png tool that allows users to
toggle between dark and light preview backgrounds (i.e. not included in PNG export)
so that they can actually see the SVGs they upload in such cases.