style: add top margin to main containers
Because the footer link pushes the main content/container up by 60px, the main
content isn't vertically centered.  The "back" button and the PageTitle component
that was implemented to emulate its design are not part of the static flow of the
page, so they do not push the content down equally. Therefore we have to add a top
margin equal to the height of the footer in order to keep the content centered both
on the main page.tsx and the tool page layout.tsx.