feat(square-tool): add cover image-fit mode to sq tool
The square image tool so far has only ever taken a source image and "contained"
it within a square representing the largest dimension of the source image.  However,
there is another way to square an image, and that is to take the smallest dimension
of the source image as the square size, and then "cover" this square with the source
image. The "object-fit" property in CSS is very similar.  I implemented this with an
instance of the OptionSelector component that toggles between the two modes, with
"contain", the status quo, as the default value.  Now, users can toggle between these
two methods of "squaring" an image at will.

I also fixed issues that were related to differences in how the HTML canvas "drawImage"
method handles SVG vs. raster data.  This was causing some SVG images to be incorrectly
positioned in "cover" mode.